Stub out useDocs hook, split out CollectionService
